+++ src/solver.jl
@@ -1276,10 +1276,9 @@ function make_dual_shadow(solver::Solver{P, U, Float64},
                           body_aero::BodyAerodynamics{P, W, Float64},
                           ::Type{TD}) where {P, U, W, TD}
     wings_d = [_wing_with_eltype(wing, TD) for wing in body_aero.wings]
-    body_aero_d = BodyAerodynamics(wings_d;
-        va = MVector{3, TD}(body_aero._va),
-        omega = MVector{3, TD}(body_aero.omega),
-    )
+    body_aero_d = BodyAerodynamics(wings_d)
+    set_va!(body_aero_d, MVector{3, TD}(body_aero._va), MVector{3, TD}(body_aero.omega);
+            reference_point=body_aero.reference_point)
     solver_d = Solver(P, U, TD;
         solver_type = solver.solver_type,
         aerodynamic_model_type = solver.aerodynamic_model_type,